Add some helper fns for generating specs from files
